+++ library/data_structures/seg_tree.hpp
@@ -10,7 +10,7 @@ inline int split(int tl, int tr) {
 	int pw2 = 1 << __lg(tr - tl);
 	return min(tl + pw2, tr - pw2 / 2);
 }
-inline long long op(long long vl, long long vr) {
+inline int64_t op(int64_t vl, int64_t vr) {
 	return vl + vr;
 }
 /**
@@ -23,7 +23,7 @@ inline long long op(long long vl, long long vr) {
  */
 struct seg_tree {
 	int n;
-	vector<long long> tree, lazy;
+	vector<int64_t> tree, lazy;
 	seg_tree(int a_n) : n(a_n), tree(2 * n), lazy(n) {}
 	seg_tree(const vector<int>& a) : n(ssize(a)), tree(2 * n), lazy(n) {
 		int pw2 = 1;
@@ -31,7 +31,7 @@ struct seg_tree {
 		for (int i = 0; i < n; i++) tree[(i + pw2) % n + n] = a[i];
 		for (int i = n - 1; i >= 1; i--) tree[i] = op(tree[2 * i], tree[2 * i + 1]);
 	}
-	inline void apply(long long change, int tl, int tr, int u) {
+	inline void apply(int64_t change, int tl, int tr, int u) {
 		tree[u] += (tr - tl) * change;
 		if (u < n) lazy[u] += change;
 	}
@@ -45,8 +45,8 @@ struct seg_tree {
 	/**
 	 * @param le,ri defines range [le, ri)
 	 */
-	void update(int le, int ri, long long change) {update(le, ri, change, 0, n, 1);}
-	void update(int le, int ri, long long change, int tl, int tr, int u) {
+	void update(int le, int ri, int64_t change) {update(le, ri, change, 0, n, 1);}
+	void update(int le, int ri, int64_t change, int tl, int tr, int u) {
 		if (ri <= tl || tr <= le) return;
 		if (le <= tl && tr <= ri) return apply(change, tl, tr, u);
 		int tm = split(tl, tr);
@@ -58,8 +58,8 @@ struct seg_tree {
 	/**
 	 * @param le,ri defines range [le, ri)
 	 */
-	long long query(int le, int ri) {return query(le, ri, 0, n, 1);}
-	long long query(int le, int ri, int tl, int tr, int u) {
+	int64_t query(int le, int ri) {return query(le, ri, 0, n, 1);}
+	int64_t query(int le, int ri, int tl, int tr, int u) {
 		if (ri <= tl || tr <= le) return 0;
 		if (le <= tl && tr <= ri) return tree[u];
 		int tm = split(tl, tr);
