+++ library/data_structures/seg_tree_uncommon/implicit.hpp
@@ -6,7 +6,7 @@
  * @endcode
  */
 template <int N> struct implicit_seg_tree {
-	using dt = array<long long, 2>; /**< min, number of mins */
+	using dt = array<int64_t, 2>; /**< min, number of mins */
 	static dt op(const dt& le, const dt& ri) {
 		if (le[0] == ri[0]) return {le[0], le[1] + ri[1]};
 		return min(le, ri);
@@ -14,14 +14,14 @@ template <int N> struct implicit_seg_tree {
 	static constexpr dt unit{LLONG_MAX, 0LL};
 	struct node {
 		dt val;
-		long long lazy = 0;
+		int64_t lazy = 0;
 		int lch = -1, rch = -1;
 	} tree[N];
 	int ptr = 0, root_l, root_r; /**< [root_l, root_r) defines range of root node; handles negatives */
 	implicit_seg_tree(int le, int ri) : root_l(le), root_r(ri) {
 		tree[ptr++].val = {0, ri - le};
 	}
-	inline void apply(long long add, int u) {
+	inline void apply(int64_t add, int u) {
 		tree[u].val[0] += add;
 		tree[u].lazy += add;
 	}
@@ -42,8 +42,8 @@ template <int N> struct implicit_seg_tree {
 	/**
 	 * @param le,ri defines range [le, ri)
 	 */
-	void update(int le, int ri, long long add) {update(le, ri, add, root_l, root_r, 0);}
-	void update(int le, int ri, long long add, int tl, int tr, int u) {
+	void update(int le, int ri, int64_t add) {update(le, ri, add, root_l, root_r, 0);}
+	void update(int le, int ri, int64_t add, int tl, int tr, int u) {
 		if (ri <= tl || tr <= le)
 			return;
 		if (le <= tl && tr <= ri)
