+++ library/data_structures/uncommon/deque_with_op.hpp
@@ -4,7 +4,7 @@
  * deque with query for operation of the deque
  * @code{.cpp}
        //deque with query for: get min and # of mins in deque
-       vector<pair<long long, int>> a; //initialize a[i].second = 1
+       vector<pair<int64_t, int>> a; //initialize a[i].second = 1
        deq dq(a, [](auto x, auto y) {
            if (x.first == y.first) return pair(x.first, x.second + y.second);
            return min(x, y);
