+++ library/data_structures/wavelet_merge/wavelet_tree.hpp
@@ -18,7 +18,7 @@ inline int split(int tl, int tr) {
 struct wavelet_tree {
 	int n, minv, maxv;
 	vector<bool_presum> bool_presums;
-	vector<vector<long long>> presums;
+	vector<vector<int64_t>> presums;
 	/**
 	 * @code{.cpp}
 	       vector<int> a;
@@ -45,7 +45,7 @@ struct wavelet_tree {
 		transform(begin(a) + le, begin(a) + ri, begin(bools), low);
 		bool_presums[u] = bool_presum(bools);
 		presums[u].resize(ri - le + 1);
-		inclusive_scan(begin(a) + le, begin(a) + ri, begin(presums[u]) + 1, plus<long long>(), 0LL);
+		inclusive_scan(begin(a) + le, begin(a) + ri, begin(presums[u]) + 1, plus<int64_t>(), 0LL);
 		int mi = stable_partition(begin(a) + le, begin(a) + ri, low) - begin(a);
 		build(a, le, mi, tl, tm, 2 * u);
 		build(a, mi, ri, tm, tr, 2 * u + 1);
@@ -73,11 +73,11 @@ struct wavelet_tree {
 	 * @time O(log(maxv - minv))
 	 * @space O(log(maxv - minv)) for recursive stack
 	 */
-	long long rect_sum(int le, int ri, int x, int y) {
+	int64_t rect_sum(int le, int ri, int x, int y) {
 		assert(0 <= le && le <= ri && ri <= n && x <= y);
 		return rect_sum_impl(le, ri, x, y, minv, maxv, 1);
 	}
-	long long rect_sum_impl(int le, int ri, int x, int y, int tl, int tr, int u) {
+	int64_t rect_sum_impl(int le, int ri, int x, int y, int tl, int tr, int u) {
 		if (y <= tl || tr <= x) return 0;
 		if (x <= tl && tr <= y) return (tr - tl == 1 ? 1LL * tl * (ri - le) : presums[u][ri] - presums[u][le]);
 		int tm = split(tl, tr), pl = bool_presums[u].popcount(le), pr = bool_presums[u].popcount(ri);
@@ -114,12 +114,12 @@ struct wavelet_tree {
 	 * @time O(log(maxv - minv))
 	 * @space O(log(maxv - minv)) for recursive stack
 	 */
-	long long kth_sum(int le, int ri, int k) {
+	int64_t kth_sum(int le, int ri, int k) {
 		assert(0 <= le && ri <= n);
 		assert(0 <= k && k <= ri - le);
 		return kth_sum_impl(le, ri, k, minv, maxv, 1);
 	}
-	long long kth_sum_impl(int le, int ri, int k, int tl, int tr, int u) {
+	int64_t kth_sum_impl(int le, int ri, int k, int tl, int tr, int u) {
 		if (tr - tl == 1) return 1LL * k * tl;
 		int tm = split(tl, tr), pl = bool_presums[u].popcount(le), pr = bool_presums[u].popcount(ri);
 		if (k <= pr - pl) return kth_sum_impl(pl, pr, k, tl, tm, 2 * u);
