+++ library/math/matrix_related/xor_basis_unordered.hpp
@@ -4,7 +4,7 @@
  * @see https://codeforces.com/blog/entry/68953 https://codeforces.com/blog/entry/100066?#comment-887897
  * @code{.cpp}
        basis<int> b;
-       basis<long long> b;
+       basis<int64_t> b;
        sort(rbegin(b.b), rend(b.b)); //needed for kth smallest walk
  * @endcode
  * notes:
