+++ tests/library_checker_aizu_tests/data_structures/disjoint_rmq_sum.test.cpp
@@ -7,10 +7,10 @@ int main() {
 	cin.tie(0)->sync_with_stdio(0);
 	int n, q;
 	cin >> n >> q;
-	vector<long long> arr(n);
+	vector<int64_t> arr(n);
 	for (int i = 0; i < n; i++)
 		cin >> arr[i];
-	disjoint_rmq rmq(arr, plus<long long> {});
+	disjoint_rmq rmq(arr, plus<int64_t> {});
 	while (q--) {
 		int l, r;
 		cin >> l >> r;
