+++ tests/library_checker_aizu_tests/graphs/edge_cd_count_paths_per_length.test.cpp
@@ -18,7 +18,7 @@ int main() {
 	{
 		edge_cd(adj, edge_cd_asserts);
 	}
-	vector<long long> cnt_len = count_paths_per_length(adj);
+	vector<int64_t> cnt_len = count_paths_per_length(adj);
 	for (int i = 1; i < n; i++)
 		cout << cnt_len[i] << " ";
 	cout << '\n';
